# docs/backend/openapi.yml
openapi: 3.1.0
info:
  title: Practice Catalog API
  description: |
    Бэкенд мобильного каталога товаров (Добрушский фарфоровый завод).

    Соглашения:
    - Деньги — целые копейки (BYN), `integer`/`int64`.
    - Даты и время — ISO-8601, UTC (`2026-08-17T12:00:00Z`).
    - Ошибки — единый формат RFC 9457 Problem Details (схема `ProblemDetails`).
    - Пагинация — `page` (0-based) + `size` (по умолчанию 20, максимум 100); конверт `{items, page, size, total, totalPages}`.
    - Идемпотентность запросов, меняющих состояние, — заголовок `X-Request-Id`.

    Подробнее: [03-api-design.md](03-api-design.md), [02-domain-model.md](02-domain-model.md).
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Локальная разработка
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/products:
    get:
      tags: [Catalog]
      summary: Поиск и список товаров
      description: |
        Поиск с фильтрами, сортировкой и пагинацией. Фильтры `inStock` и `onlyDiscounted`
        выполняются напрямую к БД (без кэша) и требуют актуальных остатков.
      security: []
      parameters:
        - name: q
          in: query
          description: Поиск по названию, артикулу, серии
          schema: { type: string, maxLength: 200 }
        - name: categoryId
          in: query
          description: Идентификатор категории (включает подкатегории)
          schema: { type: string, format: uuid }
        - name: priceFrom
          in: query
          description: Минимальная цена со скидкой, копейки
          schema: { type: integer, format: int64, minimum: 0 }
        - name: priceTo
          in: query
          description: Максимальная цена со скидкой, копейки
          schema: { type: integer, format: int64, minimum: 0 }
        - name: series
          in: query
          description: Серия/коллекция
          schema: { type: string, maxLength: 200 }
        - name: type
          in: query
          description: Тип изделия (сервиз, чашка, ваза, статуэтка...)
          schema: { type: string, maxLength: 100 }
        - name: inStock
          in: query
          description: Только товары с остатком больше 0
          schema: { type: boolean }
        - name: onlyDiscounted
          in: query
          description: Только товары со скидкой
          schema: { type: boolean }
        - name: sort
          in: query
          description: Сортировка
          schema:
            type: string
            enum: [price_asc, price_desc, rating_desc, newest, discount_desc]
            default: newest
        - name: page
          in: query
          schema: { type: integer, minimum: 0, default: 0 }
        - name: size
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      responses:
        "200":
          description: Страница товаров
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductsPage"
              examples:
                page:
                  $ref: "#/components/examples/ProductsPageExample"
        "400":
          $ref: "#/components/responses/ValidationError"
  /api/v1/products/{id}:
    get:
      tags: [Catalog]
      summary: Карточка товара
      description: "Полные данные товара: характеристики, изображения, SKU с остатками, рейтинг."
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Карточка товара
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductResponse"
              examples:
                card:
                  $ref: "#/components/examples/ProductResponseExample"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v1/admin/products:
    get:
      tags: [Admin]
      summary: Список всех товаров
      description: |
        В отличие от публичного поиска видит неактивные и мягко удалённые товары.
        `categoryId` фильтрует точным совпадением (без подкатегорий).
      parameters:
        - name: status
          in: query
          description: Фильтр по статусу товара
          schema:
            type: string
            enum: [ACTIVE, INACTIVE, DELETED]
            default: ACTIVE
        - name: q
          in: query
          description: Поиск по названию, артикулу, серии
          schema: { type: string, maxLength: 200 }
        - name: categoryId
          in: query
          schema: { type: string, format: uuid }
        - name: sort
          in: query
          schema:
            type: string
            enum: [price_asc, price_desc, rating_desc, newest, discount_desc]
            default: newest
        - name: page
          in: query
          schema: { type: integer, minimum: 0, default: 0 }
        - name: size
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      responses:
        "200":
          description: Страница товаров (любого статуса)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminProductsPage"
              examples:
                page:
                  value:
                    items:
                      - id: 7c9e6679-7425-40de-944b-e07fc1f90ae7
                        name: "Чайный сервиз «Славянский» на 6 персон"
                        article: CS-1001
                        series: Славянский
                        productType: сервиз
                        priceCents: 120000
                        discountPercent: 15
                        priceWithDiscountCents: 102000
                        mainImageUrl: http://localhost:8080/catalog-images/products/7c9e.../main.jpg
                        ratingAverage: 4.8
                        ratingCount: 23
                        isActive: false
                        deletedAt: null
                        createdAt: "2026-08-17T12:00:00Z"
                    page: 0
                    size: 20
                    total: 1
                    totalPages: 1
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [Admin]
      summary: Создать товар
      description: Цена со скидкой вычисляется автоматически.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductRequest"
            examples:
              create:
                value:
                  categoryId: a1b2c3d4-1111-4c1e-9f2c-000000000002
                  name: "Чайный сервиз «Славянский» на 6 персон"
                  article: CS-1001
                  description: Классический чайный сервиз из твёрдого фарфора
                  series: Славянский
                  productType: сервиз
                  decor: синий кобальт
                  material: твёрдый фарфор
                  capacityMl: 220
                  weightG: 3400
                  dimensions: "310x210x180 мм"
                  countryOfOrigin: Беларусь
                  barcode: "4810123000014"
                  priceCents: 120000
                  discountPercent: 15
                  isActive: true
      responses:
        "201":
          description: Товар создан
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/admin/products/{id}:
    get:
      tags: [Admin]
      summary: Карточка товара в любом статусе
      description: |
        Полная карточка, включая неактивные и мягко удалённые товары (публичный GET отдаёт 404).
        Возвращает флаги isActive/deletedAt и все SKU, включая неактивные —
        для просмотра перед реактивацией.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Карточка товара
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminProductCard"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Admin]
      summary: Обновить товар
      description: Частичное обновление; пересчёт цены со скидкой; событие product.updated.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdateRequest"
      responses:
        "200":
          description: Обновлённый товар
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
    delete:
      tags: [Admin]
      summary: Удалить товар (soft delete)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204":
          description: Товар удалён
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/admin/products/{id}/skus:
    post:
      tags: [Admin]
      summary: Создать вариант товара (SKU)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SkuRequest"
            examples:
              create:
                value:
                  name: На 6 персон
                  article: CS-1001-6
                  priceCents: 120000
                  discountPercent: 15
                  stockQty: 14
      responses:
        "201":
          description: SKU создан
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SkuResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/admin/skus/{id}:
    patch:
      tags: [Admin]
      summary: Обновить SKU (цена, скидка, остаток)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SkuUpdateRequest"
            examples:
              update:
                value: { priceCents: 115000, discountPercent: 10, stockQty: 20 }
      responses:
        "200":
          description: Обновлённый SKU
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SkuResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Admin]
      summary: Деактивировать SKU
      description: SKU деактивируется (не удаляется); товар без активных SKU не продаётся.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204":
          description: SKU деактивирован
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/admin/products/{id}/images:
    post:
      tags: [Admin]
      summary: Загрузить изображение товара
      description: |
        Multipart-загрузка. Форматы JPEG/PNG/WebP, до 10 МБ.
        Первое загруженное изображение автоматически становится главным.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: Файл изображения (JPEG/PNG/WebP, до 10 МБ)
                position:
                  type: integer
                  minimum: 0
                  description: Порядковый номер (по умолчанию — конец)
                isMain:
                  type: boolean
                  description: Сделать главным изображением
            required: [file]
      responses:
        "201":
          description: Изображение загружено
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductImageResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/admin/images/{id}:
    patch:
      tags: [Admin]
      summary: Изменить позицию/главное изображение
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImagePositionRequest"
            examples:
              setMain:
                value: { position: 0, isMain: true }
      responses:
        "200":
          description: Изображение обновлено
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductImageResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Admin]
      summary: Удалить изображение
      description: Удаляет запись и объект из MinIO.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204":
          description: Изображение удалено
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access-токен, выданный через /api/v1/auth/login или /api/v1/auth/refresh
  examples:
...
    ProductsPageExample:
      value:
        items:
          - id: 7c9e6679-7425-40de-944b-e07fc1f90ae7
            name: "Чайный сервиз «Славянский» на 6 персон"
            article: CS-1001
            series: Славянский
            productType: сервиз
            priceCents: 120000
            discountPercent: 15
            priceWithDiscountCents: 102000
            mainImageUrl: http://localhost:8080/catalog-images/products/7c9e.../main.jpg
            ratingAverage: 4.8
            ratingCount: 23
            inStock: true
        page: 0
        size: 10
        total: 5
        totalPages: 1
    ProductResponseExample:
      value:
        id: 7c9e6679-7425-40de-944b-e07fc1f90ae7
        categoryId: a1b2c3d4-1111-4c1e-9f2c-000000000002
        name: "Чайный сервиз «Славянский» на 6 персон"
        article: CS-1001
        description: Классический чайный сервиз из твёрдого фарфора
        series: Славянский
        productType: сервиз
        decor: синий кобальт
        material: твёрдый фарфор
        capacityMl: 220
        weightG: 3400
        dimensions: "310x210x180 мм"
        countryOfOrigin: Беларусь
        barcode: "4810123000014"
        priceCents: 120000
        discountPercent: 15
        priceWithDiscountCents: 102000
        ratingAverage: 4.8
        ratingCount: 23
        images:
          - id: i1
            url: http://localhost:8080/catalog-images/products/7c9e.../main.jpg
            position: 0
            isMain: true
        skus:
          - id: s1
            name: На 6 персон
            article: CS-1001-6
            priceCents: 120000
            priceWithDiscountCents: 102000
            stockQty: 14
            isActive: true
...
  responses:
    ValidationError:
      description: Ошибка валидации полей запроса
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationProblemDetails"
    Unauthorized:
      description: Отсутствует, просрочен или невалиден токен
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    Forbidden:
      description: Недостаточно прав (роль)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    NotFound:
      description: Ресурс не найден
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    Conflict:
      description: Конфликт состояния (409)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
...
  schemas:
...
    ValidationProblemDetails:
      allOf:
        - $ref: "#/components/schemas/ProblemDetails"
        - type: object
          properties:
            errors:
              type: object
              description: "Карта ошибок валидации: поле → список сообщений"
              additionalProperties:
                type: array
                items:
                  type: string
              example:
                email: ["должен быть валидным email"]
                password: ["длина от 8 до 72 символов"]
    PageInfo:
      type: object
      description: Метаданные пагинации
      required: [page, size, total, totalPages]
      properties:
        page:
          type: integer
          minimum: 0
        size:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
...
    ProductsPage:
      allOf:
        - $ref: "#/components/schemas/PageInfo"
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/ProductListItem"
    ProductImageResponse:
      type: object
      required: [id, url, position, isMain]
      properties:
        id: { type: string, format: uuid }
        url: { type: string, format: uri, description: Публичный URL изображения }
        position: { type: integer, minimum: 0 }
        isMain: { type: boolean }
    SkuResponse:
      type: object
      required: [id, name, article, priceCents, priceWithDiscountCents, stockQty, isActive]
      properties:
        id: { type: string, format: uuid }
        name: { type: string, maxLength: 200 }
        article: { type: string, maxLength: 64 }
        priceCents: { type: integer, format: int64, nullable: true, description: Собственная цена SKU, копейки BYN; null — цена товара }
        priceWithDiscountCents: { type: integer, format: int64, nullable: true }
        stockQty: { type: integer, minimum: 0, description: Остаток, всегда из БД }
        isActive: { type: boolean }
    ProductResponse:
      type: object
      description: Карточка товара
      required:
        [id, categoryId, name, article, series, productType, priceCents, priceWithDiscountCents, ratingAverage, ratingCount, images, skus]
      properties:
        id: { type: string, format: uuid }
        categoryId: { type: string, format: uuid }
        name: { type: string, maxLength: 300 }
        article: { type: string, maxLength: 64 }
        description: { type: string, nullable: true }
        series: { type: string, maxLength: 200 }
        productType: { type: string, maxLength: 100 }
        decor: { type: string, maxLength: 200, nullable: true }
        material: { type: string, maxLength: 100, nullable: true }
        capacityMl: { type: integer, nullable: true, description: Вместимость, мл }
        weightG: { type: integer, nullable: true, description: Вес, г }
        dimensions: { type: string, maxLength: 100, nullable: true, description: "Габариты ДхШхВ" }
        countryOfOrigin: { type: string, maxLength: 100, nullable: true }
        barcode: { type: string, maxLength: 32, nullable: true }
        priceCents: { type: integer, format: int64, description: Цена, копейки BYN }
        discountPercent: { type: integer, minimum: 0, maximum: 100, nullable: true }
        priceWithDiscountCents: { type: integer, format: int64, description: Цена со скидкой, копейки BYN }
        ratingAverage: { type: number, format: double, minimum: 0, maximum: 5 }
        ratingCount: { type: integer, minimum: 0 }
        images:
          type: array
          items:
            $ref: "#/components/schemas/ProductImageResponse"
        skus:
          type: array
          items:
            $ref: "#/components/schemas/SkuResponse"
...
    ProductRequest:
      type: object
      description: Создание товара; цена со скидкой вычисляется автоматически
      required: [categoryId, name, article, productType, priceCents]
      properties:
        categoryId: { type: string, format: uuid }
        name: { type: string, maxLength: 300 }
        article: { type: string, maxLength: 64, description: Уникален среди активных }
        description: { type: string, nullable: true }
        series: { type: string, maxLength: 200, nullable: true }
        productType: { type: string, maxLength: 100 }
        decor: { type: string, maxLength: 200, nullable: true }
        material: { type: string, maxLength: 100, nullable: true }
        capacityMl: { type: integer, nullable: true }
        weightG: { type: integer, nullable: true }
        dimensions: { type: string, maxLength: 100, nullable: true }
        countryOfOrigin: { type: string, maxLength: 100, nullable: true }
        barcode: { type: string, maxLength: 32, nullable: true }
        priceCents: { type: integer, format: int64, minimum: 0, description: Цена, копейки BYN }
        discountPercent: { type: integer, minimum: 0, maximum: 100, default: 0 }
        isActive: { type: boolean, default: true }
    ProductUpdateRequest:
      type: object
      description: Частичное обновление товара; все поля опциональны
      properties:
        categoryId: { type: string, format: uuid }
        name: { type: string, maxLength: 300 }
        description: { type: string, nullable: true }
        series: { type: string, maxLength: 200, nullable: true }
        productType: { type: string, maxLength: 100 }
        decor: { type: string, maxLength: 200, nullable: true }
        material: { type: string, maxLength: 100, nullable: true }
        capacityMl: { type: integer, nullable: true }
        weightG: { type: integer, nullable: true }
        dimensions: { type: string, maxLength: 100, nullable: true }
        countryOfOrigin: { type: string, maxLength: 100, nullable: true }
        barcode: { type: string, maxLength: 32, nullable: true }
        priceCents: { type: integer, format: int64, minimum: 0 }
        discountPercent: { type: integer, minimum: 0, maximum: 100 }
        isActive: { type: boolean }
      minProperties: 1
    SkuRequest:
      type: object
      required: [name, article, stockQty]
      properties:
        name: { type: string, maxLength: 200 }
        article: { type: string, maxLength: 64, description: Уникален среди активных }
        priceCents: { type: integer, format: int64, minimum: 0, nullable: true, description: null — цена товара }
        discountPercent: { type: integer, minimum: 0, maximum: 100, default: 0 }
        stockQty: { type: integer, minimum: 0 }
    SkuUpdateRequest:
      type: object
      description: Частичное обновление SKU
      properties:
        name: { type: string, maxLength: 200 }
        priceCents: { type: integer, format: int64, minimum: 0, nullable: true }
        discountPercent: { type: integer, minimum: 0, maximum: 100 }
        stockQty: { type: integer, minimum: 0 }
      minProperties: 1
    ImagePositionRequest:
      type: object
      properties:
        position: { type: integer, minimum: 0 }
        isMain: { type: boolean }
      minProperties: 1
...
    AdminProductsPage:
      allOf:
        - $ref: "#/components/schemas/PageInfo"
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/AdminProductListItem"
    AdminProductCard:
      allOf:
        - $ref: "#/components/schemas/ProductResponse"
        - type: object
          required: [isActive]
          properties:
            isActive:
              type: boolean
              description: false — товар скрыт из публичного каталога
            deletedAt:
              type: string
              format: date-time
              nullable: true
              description: Момент мягкого удаления; null — не удалён
...